Time: 54 ms (8.2%), Space: 48.4 MB (46.21%) - LeetHub

+++ 1002-maximum-width-ramp/1002-maximum-width-ramp.cpp
@@ -24,8 +24,8 @@ class Solution {
 public:
     int maxWidthRamp(vector<int>& nums) {
         int  maxWidth = 0;
-        unsigned short i;
-        stack<unsigned short>stk;
+        int i;
+        stack<int>stk;
         stk.push(0);
         for ( i = 1 ; i < nums.size() ; i++ )
         {
@@ -33,7 +33,7 @@ public:
                 stk.push(i);
         }
 
-        for (unsigned short j = nums.size() - 1; j >= 0; j--)
+        for (int j = nums.size() - 1; j >= 0; j--)
         {
             
             while (nums[j] >= nums[stk.top()])
